+++ src/index.ts
@@ -1,3 +1,4 @@
 export { default } from './deployStack';
 export { default as createTableWaiter }  from './createTableWaiter';
+export { default as createEventLogWaiter }  from './createEventLogWaiter';
 export * from './deployStack';